Refactor: Move HealthCheck entity and related repositories to the data package
This commit moves the `HealthCheck` entity and its related repositories (`HealthCheckRepository`, `ChildRepository`, `TaskRepository`) from the `presentation.data` package to the `data` package.
This change improves the package structure by placing data-related entities and their repositories in a more appropriate location.
The `HealthScreen` and `HealthViewModel` are updated to reflect this change.