Refactor: Move entity classes to data layer
This commit moves the `Meal` and `Medication` entity classes from the `presentation.data.local.entity` package to the `data.local.entity` package. It also updates the import statements in `MealScreen`, `MedicationScreen`, and `MealScreenViewModel` accordingly to reflect the changes in the package structure. This improves the separation of concerns and code organization by placing data-related entities in the `data` layer.